+++ src/com/mdoyle/Planetarium/DataTypes/Quantity.java
@@ -1,19 +1,21 @@
 package com.mdoyle.Planetarium.DataTypes;
 
+import java.math.BigDecimal;
+
 public abstract class Quantity {
 
-	private double value;
+	private BigDecimal value;
 	private String unit;
 	
-	public Quantity(double value) {
+	public Quantity(BigDecimal value) {
 		this.value = value;
 		this.unit = getUnit();
 	}
 	
-	public double getValue() {
+	public BigDecimal getValue() {
 		return value;
 	}
-	public void setValue(double value) {
+	public void setValue(BigDecimal value) {
 		this.value = value;
 	}
 	public abstract String getUnit();
